pyproject.toml
[project]
name = "nas-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.42.0",
    "ipykernel>=7.1.0",
    "numpy<2",
    "pandas>=2.3.3",
    "pydantic>=2.12.0",
    "tabulate>=0.9.0",
    "torch>=2.8.0",
]
